Update musicrecplus.py
So I found that in the findBestUser() function, the variable 'bestUser' isn't assigning to user as stated in the code. In fact, it will remain the 'None' type it started out as. If you change the last line from 'return bestUser' to 'return user', the function will now in fact return everyone's info, though I am not sure if that is what we want. Just letting you guys know what I've observed after playing with the code.